fix(cache): env-driven persistent-cache min_compile_time (default 0, CI sets 1) + test hygiene (#1260)
* fix(cache): env-driven persistent-cache min_compile_time + test cache hygiene

Make CI benefit from the shared /xla-cache (GCS) persistent compilation cache
instead of regressing on it.

scheduler.py:
- The persistent-cache compile-time threshold
  (jax_persistent_cache_min_compile_time_secs) is now env-driven: default 0
  (cache every compile) for local/dev, and when
  JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS is set, defer to JAX's own parsing
  so value validation matches upstream JAX. Previously it was hardcoded to 0.
- Log the resolved persistent-cache state (enabled / disabled / not-configured,
  plus the effective min_compile_time) at startup.

CI workflows (pr-test, nightly-test-daily, multi-host-test):
- Set JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS=1 wherever
  JAX_COMPILATION_CACHE_DIR is configured, so sub-1s compilations are not written
  to the GCS-backed cache (~3x fewer small-file writes per cold run); local/dev
  keeps the default 0.

Tests:
- Stop overriding JAX_COMPILATION_CACHE_DIR in individual tests so they share the
  CI-provided cache dir (cache content is content-addressed, so sharing is safe).
- Enable precompile + cache-miss checking in the w8 block-quant tests.

Closes #1263


* ci(test): rebalance e2e-test-tpu-v6e-4 shard weights from measured runtimes

estimated_time weights were badly miscalibrated (wan2.1/flux weighted 30 but
~6/3 min actual; test_retract_decode weighted 12 but ~18 min, test_logprobs_dp
8 but ~13 min cold). auto_partition balanced by weight while real runtimes were
lopsided — e2e-test-4-tpu shard 2 ran ~29 min warm vs ~11-12 for shards 0/1.

Recalibrate from measured cold-cache per-file runtimes (with ~20% headroom on
the longer tests). With 3 shards the slowest shard drops ~29->21 min warm
(~46->28 min cold), cutting the PR Test critical path accordingly.


* test(flashattention): split monolithic kernel test into mha/gqa/misc

The single test_flashattention.py (26 test methods, ~15 min as one TestFile)
was the serial floor of unit-test-tpu-v6e-1. Split the shared setUp/run_test
workhorse into flashattention_common.AttentionTestBase and partition the 26
methods into three files (mha=10, gqa=10, misc=6) so run_suite.py can spread
them across shards instead of running them as one indivisible block.

Logic is moved verbatim; no test behavior changes.


* ci(test): bump unit-test-tpu-v6e-1 to 5 shards

With test_flashattention split into three files, raise unit-test-1-tpu from
3 to 5 shards to parallelize the stage-1 floor. v6e-1 ARC runners have spare
capacity (max 10, well above current usage).


* ci(test): recalibrate e2e-test-tpu-v6e-4 weights from warm runtimes

The earlier weights were miscalibrated against cold/stale runtimes, leaving
test_logprobs_dp at 17 (real warm: 2.2min) and test_retract_decode at 22
(real: 16min). The greedy bin-packer split those two apart as if both were
large, producing badly skewed shards: 23.3 / 9.8 / 19.1 min (spread 13.5min).

Recalibrate every file's estimated_time from this run's measured warm
elapsed. Verified via run_suite.auto_partition: shards rebalance to
17.4 / 16.6 / 18.2 min (spread 1.6min), so the slowest e2e-4-tpu shard drops
23.3 -> 18.2 min. This is near the ~17.4min floor set by the single
test_retract_decode.py file (not split here).


---------
