+++ .github/workflows/multi-host-test.yml
@@ -360,6 +360,8 @@ jobs:
                     value: "/models/model_scope"
                   - name: JAX_COMPILATION_CACHE_DIR
                     value: "/xla-cache"
+                  - name: JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS
+                    value: "1"
                   - name: HF_HUB_DOWNLOAD_TIMEOUT
                     value: "600"
                   - name: REPO_REF
