+++ .github/workflows/pr-test.yml
@@ -28,7 +28,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        part: [0, 1, 2]
+        part: [0, 1, 2, 3, 4]
     steps:
       - name: Checkout code
         uses: actions/checkout@v5
@@ -40,9 +40,10 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
         run: |
-          python test/srt/run_suite.py --suite unit-test-tpu-v6e-1 --auto-partition-id ${{ matrix.part }} --auto-partition-size 3
+          python test/srt/run_suite.py --suite unit-test-tpu-v6e-1 --auto-partition-id ${{ matrix.part }} --auto-partition-size 5
 
   # =============================================== Stage 2: multi-chip + e2e + accuracy ============================
   unit-test-4-tpu:
@@ -64,6 +65,7 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
         run: |
           bash scripts/killall_sglang.sh
@@ -108,6 +110,7 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
           HF_TOKEN: ${{ secrets.HF_TOKEN }}
           HF_HUB_DOWNLOAD_TIMEOUT: 300
@@ -137,6 +140,7 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
           HF_TOKEN: ${{ secrets.HF_TOKEN }}
           HF_HUB_DOWNLOAD_TIMEOUT: 300
@@ -164,6 +168,7 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
           HF_TOKEN: ${{ secrets.HF_TOKEN }}
           HF_HUB_DOWNLOAD_TIMEOUT: 300
@@ -192,6 +197,7 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
           HF_TOKEN: ${{ secrets.HF_TOKEN }}
           HF_HUB_DOWNLOAD_TIMEOUT: 300
@@ -219,6 +225,7 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
           HF_TOKEN: ${{ secrets.HF_TOKEN }}
           HF_HUB_DOWNLOAD_TIMEOUT: 300
@@ -246,6 +253,7 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
           HF_TOKEN: ${{ secrets.HF_TOKEN }}
           HF_HUB_DOWNLOAD_TIMEOUT: 300
@@ -274,6 +282,7 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
           HF_TOKEN: ${{ secrets.HF_TOKEN }}
           HF_HUB_DOWNLOAD_TIMEOUT: 300
@@ -301,6 +310,7 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
           HF_TOKEN: ${{ secrets.HF_TOKEN }}
           HF_HUB_DOWNLOAD_TIMEOUT: 300
@@ -326,6 +336,7 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
           HF_TOKEN: ${{ secrets.HF_TOKEN }}
           HF_HUB_DOWNLOAD_TIMEOUT: 300
@@ -354,6 +365,7 @@ jobs:
         env:
           SGLANG_JAX_IS_IN_CI: true
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
           HF_TOKEN: ${{ secrets.HF_TOKEN }}
           HF_HUB_DOWNLOAD_TIMEOUT: 300
