+++ .github/workflows/weekly-test.yml
@@ -29,6 +29,7 @@ jobs:
     env:
       TZ: Asia/Shanghai
       JAX_COMPILATION_CACHE_DIR: /xla-cache
+      JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
       SGLANG_JAX_MODEL_CACHE: /models/model_scope
     steps:
       - name: Checkout code
@@ -93,6 +94,7 @@ jobs:
     env:
       TZ: Asia/Shanghai
       JAX_COMPILATION_CACHE_DIR: /xla-cache
+      JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
       SGLANG_JAX_MODEL_CACHE: /models/model_scope
     steps:
       - name: Checkout code
@@ -131,6 +133,7 @@ jobs:
     env:
       TZ: Asia/Shanghai
       JAX_COMPILATION_CACHE_DIR: /xla-cache
+      JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
       SGLANG_JAX_MODEL_CACHE: /models/model_scope
     steps:
       - name: Checkout code
@@ -174,6 +177,7 @@ jobs:
     env:
       TZ: Asia/Shanghai
       JAX_COMPILATION_CACHE_DIR: /xla-cache
+      JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
       SGLANG_JAX_MODEL_CACHE: /models/model_scope
     steps:
       - name: Checkout code
@@ -239,6 +243,7 @@ jobs:
       env:
         TZ: Asia/Shanghai
         JAX_COMPILATION_CACHE_DIR: /xla-cache
+        JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
         SGLANG_JAX_MODEL_CACHE: /models/model_scope
       steps:
         - name: Checkout code
@@ -282,6 +287,7 @@ jobs:
         env:
           TZ: Asia/Shanghai
           JAX_COMPILATION_CACHE_DIR: /xla-cache
+          JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS: "1"
           SGLANG_JAX_MODEL_CACHE: /models/model_scope
         steps:
           - name: Checkout code
