+++ python/sgl_jax/srt/managers/scheduler.py
@@ -149,6 +149,8 @@ class Scheduler(
         # set jit cache
         jit_cache_dir = os.getenv("JAX_COMPILATION_CACHE_DIR", None)
         device_indexes = server_args.device_indexes
+        # Report the effective persistent-cache state after resolving overrides below.
+        cache_status = None
         # libtpu (tpu-v6e + libtpu 0.0.30) crashes during JAX persistent
         # compilation-cache use when the device subset does not start at
         # device 0 (e.g. device_indexes=[2, 3]). Disable the cache for
@@ -171,14 +173,31 @@ class Scheduler(
             from jax.experimental.compilation_cache import compilation_cache as cc
 
             cc.reset_cache()
+            cache_status = (
+                f"disabled for non-zero-base device subset: device_indexes={device_indexes}"
+            )
         if jit_cache_dir is not None:
             jax.config.update("jax_compilation_cache_dir", jit_cache_dir)
+            # Default the compile-time write threshold to 0 (cache every compile) for
+            # local/dev. When JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS is set (CI sets
+            # it to 1 to skip tiny entries and cut small-file GCS writes), defer to JAX's
+            # own parsing so the behavior — including validation of bad values — matches
+            # upstream JAX.
+            if "JAX_PERSISTENT_CACHE_MIN_COMPILE_TIME_SECS" not in os.environ:
+                jax.config.update("jax_persistent_cache_min_compile_time_secs", 0)
+            # Disable the size gate; the compile-time threshold still controls writes.
             jax.config.update("jax_persistent_cache_min_entry_size_bytes", -1)
-            jax.config.update("jax_persistent_cache_min_compile_time_secs", 0)
+            # Include XLA sub-caches such as kernel/autotune data.
             jax.config.update("jax_persistent_cache_enable_xla_caches", "all")
             from jax.experimental.compilation_cache import compilation_cache as cc
 
             cc.set_cache_dir(jit_cache_dir)
+            min_compile_time = jax.config.jax_persistent_cache_min_compile_time_secs
+            cache_status = f"enabled, dir={jit_cache_dir}, min_compile_time={min_compile_time}s"
+
+        if cache_status is None:
+            cache_status = "not configured (JAX_COMPILATION_CACHE_DIR unset)"
+        logger.info("XLA persistent compilation cache: %s", cache_status)
 
         # Parse args
         self.server_args = server_args
