+++ test/srt/lora/test_static_lora.py
@@ -22,9 +22,7 @@ class TestStaticLoRA(CustomTestCase):
         print("=================== test_diff_after_apply_dummy =======================")
         model_path = QWEN3_4B
         lora_target_modules = ["gate_proj"]
-        import os
 
-        os.environ["JAX_COMPILATION_CACHE_DIR"] = "/tmp/jit_cache"
         engine = Engine(
             model_path=model_path,
             trust_remote_code=True,
