+++ test/srt/run_suite.py
@@ -254,7 +254,9 @@ suites = {
     "unit-test-tpu-v6e-1": [
         TestFile("python/sgl_jax/test/kernels/quantized_linear_test.py", 0.3, runner="pytest"),
         TestFile("python/sgl_jax/test/kernels/moe_block_quant_test.py", 0.2, runner="pytest"),
-        TestFile("python/sgl_jax/test/test_flashattention.py", 15),
+        TestFile("python/sgl_jax/test/test_flashattention_mha.py", 11),
+        TestFile("python/sgl_jax/test/test_flashattention_gqa.py", 11),
+        TestFile("python/sgl_jax/test/test_flashattention_misc.py", 7),
         TestFile("python/sgl_jax/test/test_mla_attention.py", 2.5),
         TestFile("python/sgl_jax/test/test_moe_topk.py", 0.3),
         TestFile("python/sgl_jax/test/kernels/fused_moe_v1_test.py", 9),
@@ -367,20 +369,20 @@ suites = {
         TestFile("test/srt/lora/test_static_lora.py", 10),
     ],
     "e2e-test-tpu-v6e-4": [
-        TestFile("test/srt/openai_server/basic/test_tool_calls.py", 3),
-        TestFile("test/srt/test_features.py", 10),
-        TestFile("test/srt/test_chunked_prefill_size.py", 5),
+        TestFile("test/srt/openai_server/basic/test_tool_calls.py", 2),
+        TestFile("test/srt/test_features.py", 3),
+        TestFile("test/srt/test_chunked_prefill_size.py", 2),
         # TestFile("test/srt/test_sliding_window_attention.py", 30), # add after gpt-oss supported
-        TestFile("test/srt/test_logprobs_dp.py", 8),
-        TestFile("test/srt/test_model_loader.py", 5),
-        TestFile("test/srt/test_deepseek_v2_lite_models.py", 10),
-        TestFile("test/srt/quantization/test_w8_quantization.py", 10),
+        TestFile("test/srt/test_logprobs_dp.py", 3),
+        TestFile("test/srt/test_model_loader.py", 1),
+        TestFile("test/srt/test_deepseek_v2_lite_models.py", 5),
+        TestFile("test/srt/quantization/test_w8_quantization.py", 4),
         TestFile(
             "test/srt/quantization/test_w8_block_dynamic_quantization.py",
-            8,
+            2,
             runner="pytest",
         ),
-        TestFile("test/srt/quantization/test_w8_moe_block_linear_channel_quantization.py", 15),
+        TestFile("test/srt/quantization/test_w8_moe_block_linear_channel_quantization.py", 3),
         # TestFile("test/srt/test_engine_determine_generation.py", 5),
         # ^ Disabled in DP merge: asserts bit-exact equivalence between baseline,
         #   retract, and abort+regenerate generation paths under temperature=0.
@@ -389,14 +391,14 @@ suites = {
         #   so re-prefilling [prompt + N generated tokens] after retract no longer reproduces
         #   the original decode-step logits bit-exactly, causing argmax to diverge.
         #   Tracked separately; not a regression introduced by this PR.
-        TestFile("test/srt/test_engine_flush_cache.py", 5),
-        TestFile("test/srt/test_engine_pause_continue.py", 6),
-        TestFile("test/srt/test_server_pause_continue.py", 6),
-        TestFile("test/srt/test_retract_decode.py", 12),
-        TestFile("test/srt/rl/test_return_routed_experts.py", 5),
-        TestFile("test/srt/rl/test_multi_engines_in_one_process.py", 5),
-        TestFile("test/srt/multimodal/test_wan2_1_models.py", 30),
-        TestFile("test/srt/multimodal/test_flux1_dev_models.py", 30),
+        TestFile("test/srt/test_engine_flush_cache.py", 1),
+        TestFile("test/srt/test_engine_pause_continue.py", 2),
+        TestFile("test/srt/test_server_pause_continue.py", 3),
+        TestFile("test/srt/test_retract_decode.py", 17),
+        TestFile("test/srt/rl/test_return_routed_experts.py", 2),
+        TestFile("test/srt/rl/test_multi_engines_in_one_process.py", 1),
+        TestFile("test/srt/multimodal/test_wan2_1_models.py", 4),
+        TestFile("test/srt/multimodal/test_flux1_dev_models.py", 2),
     ],
 }
