+++ test/srt/test_bench_serving_dense.py
@@ -51,9 +51,6 @@ class TestBenchServingDense(CustomTestCase):
                 "128",
                 "--disable-radix-cache",
             ],
-            env={
-                "JAX_COMPILATION_CACHE_DIR": "/tmp/jax_compilation_cache",
-            },
         )
 
     @classmethod
