+++ test/srt/test_chunked_prefill_size.py
@@ -45,9 +45,6 @@ class TestChunkedPrefillSize(CustomTestCase):
                 "--disable-radix-cache",
                 "--enable-mixed-chunk",
             ],
-            env={
-                "JAX_COMPILATION_CACHE_DIR": "/tmp/jax_compilation_cache",
-            },
         )
 
     @classmethod
