+++ test/srt/test_eval_accuracy_large.py
@@ -61,9 +61,6 @@ class TestEvalAccuracyLarge(CustomTestCase):
                 "--page-size",
                 "64",
             ],
-            env={
-                "JAX_COMPILATION_CACHE_DIR": "/tmp/jax_compilation_cache",
-            },
         )
 
     @classmethod
