+++ test/srt/test_penalty.py
@@ -59,9 +59,6 @@ class TestPenalty(CustomTestCase):
                 "--max-running-requests",
                 "64",
             ],
-            env={
-                "JAX_COMPILATION_CACHE_DIR": "/tmp/jax_compilation_cache",
-            },
         )
 
     @classmethod
