+++ test/srt/test_server_pause_continue.py
@@ -79,9 +79,6 @@ class TestPauseContinueGeneration(CustomTestCase):
                 "--max-running-requests",
                 "64",
             ],
-            env={
-                "JAX_COMPILATION_CACHE_DIR": "/tmp/jax_compilation_cache",
-            },
         )
 
     @classmethod
@@ -543,9 +540,6 @@ class TestPauseContinueGenerationNoOverlap(CustomTestCase):
                 "64",
                 "--disable-overlap-schedule",
             ],
-            env={
-                "JAX_COMPILATION_CACHE_DIR": "/tmp/jax_compilation_cache",
-            },
         )
 
     @classmethod
