+++ test/srt/test_speculative_decoding.py
@@ -69,9 +69,6 @@ class TestSpeculativeDecoding(CustomTestCase):
                 "--tp-size",
                 "4",
             ],
-            env={
-                "JAX_COMPILATION_CACHE_DIR": "/tmp/jax_compilation_cache",
-            },
         )
 
     @classmethod
